+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ## Added
 
+- Allow to bulk sign contracts by training
 - Filtering orders by organization, product, and course product relation
 - Allow to set CSRF_COOKIE_DOMAIN through env vars
 - Dedicated storage for easy_thumbnail using boto3